Update print numberz

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -6,7 +6,8 @@
 */
 int main(void)
 {
-	long n = 0;
+	int n = 0;
+
 	while (n < 10)
 	{
 		putchar((n % 10) + '0');
